+++ dalesbred/src/main/java/org/dalesbred/transaction/TransactionSettings.java
@@ -23,7 +23,6 @@
 package org.dalesbred.transaction;
 
 import org.dalesbred.Database;
-import org.dalesbred.annotation.DalesbredTransactional;
 import org.jetbrains.annotations.NotNull;
 
 import static java.util.Objects.requireNonNull;
@@ -85,17 +84,5 @@ public final class TransactionSettings {
     @Override
     public String toString() {
         return "[propagation=" + propagation + ", isolation=" + isolation + ", retries=" + retries + ']';
-    }
-
-    /**
-     * Constructs TransactionSettings from given annotation.
-     */
-    @NotNull
-    public static TransactionSettings fromAnnotation(@NotNull DalesbredTransactional transactional) {
-        TransactionSettings settings = new TransactionSettings();
-        settings.setIsolation(transactional.isolation());
-        settings.setPropagation(transactional.propagation());
-        settings.setRetries(transactional.retries());
-        return settings;
     }
 }